exampleapp/requirements.txt
Django==2.0.1
django-rest-mock==0.2.1
djangorestframework==3.7.7